+++ packages/kokoas-server/src/handleRequest/reqCostManagement/getCostMgtDataByProjId.ts
@@ -1,4 +1,4 @@
-import { CostManagement, summarizeSuppliers } from './summarizeSuppliers';
+import { summarizeSuppliers } from './summarizeSuppliers';
 import { 
   getAndpadPaymentsBySystemId, 
   getAndpadProcurementByAndpadProjId, 
@@ -11,33 +11,9 @@ import { calcProfitability } from 'api-kintone/src/andpadProcurement/calculation
 import { getOrderByProjId } from 'api-andpad';
 import { getAgentNamesByType as custGetAgentsNamesByType } from 'api-kintone/src/custgroups/helpers/getAgentNamesByType';
 import { getAgentNamesByType as projGetAgentNamesByType } from 'api-kintone/src/projects/helpers/getAgentNamesByType';
+import type { GetCostManagement } from 'types';
 
 
-export interface GetCostManagement {
-  projNum: string,
-  projName: string,
-  custGroupName: string,
-  受注金額_税抜: number,
-  追加金額_税抜: number,
-  発注金額_税抜: number,
-  支払金額_税抜: number,
-  予定利益率: number,
-  予定利益額: number,
-  実利益率: number,
-  実利益額: number,
-  利益配分_夢てつ: number,
-  利益配分_ここすも: number,
-  実利益税抜_夢てつ: number,
-  実利益税抜_ここすも: number,
-  受注額計_税込: number,
-  受注額計_税抜: number,
-  入金額: number,
-  未入金: number,
-  夢てつ営業: string,
-  ここすも営業: string,
-  ここすも工事: string,
-  発注情報詳細: CostManagement,
-}
 
 /**
  * 必要なファイルを取得する
