+++ packages/types/src/common/index.ts
@@ -1,5 +1,7 @@
 export * from './andpad.member';
+export * from './andpad.order';
 export * from './auth';
+export * from './costManagement';
 export * from './customers';
 export * from './docusign';
 export * from './domEvents';
